update flag library build step
